test(game): 게임 테스트 15개를 :app에서 :game 모듈로 이동
이슈 #1296 :game 테스트 분리. 6개 게임의 단위·통합 테스트와 minigame
이벤트 리스너 테스트를 :game으로 이전하고, extends 대상을 app 전용
베이스에서 GameModule* 베이스 클래스로 교체한다.

:app에는 멀티 모듈 ArchUnit 테스트(GameArchitectureTest,
GameLayerArchitectureTest)와 cross-module 통합 테스트만 남는다.
