+++ backend/game/src/test/java/coffeeshout/speedtouch/application/SpeedTouchGameProgressHandlerTest.java
@@ -6,7 +6,7 @@ import static org.mockito.Mockito.atLeastOnce;
 import static org.mockito.Mockito.verify;
 
 import coffeeshout.fixture.RoomFixture;
-import coffeeshout.support.app.ServiceTest;
+import coffeeshout.GameModuleServiceTest;
 import coffeeshout.room.domain.Room;
 import coffeeshout.room.domain.player.PlayerName;
 import coffeeshout.room.domain.repository.RoomRepository;
@@ -18,7 +18,7 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
-class SpeedTouchGameProgressHandlerTest extends ServiceTest {
+class SpeedTouchGameProgressHandlerTest extends GameModuleServiceTest {
 
     @Autowired
     private RoomRepository roomRepository;
